Get create sarcophagus working with diamond pattern
