--- contracts/diamond/scripts/deploy-diamond.js
@@ -1,77 +1,0 @@
-// Unmodified from copied version
-// https://github.com/mudgen/diamond-1-hardhat/blob/main/scripts/deploy.js
-
-/* global ethers */
-/* eslint prefer-const: "off" */
-/* eslint no-process-exit: "off" */
-
-// TODO: Create a hardhat task to run tests without logs
-// Comment/uncomment to toggle console logs in this file
-// console.log = () => {};
-
-const { getSelectors, FacetCutAction } = require("./libraries/diamond.js");
-
-async function deployDiamond() {
-  const accounts = await ethers.getSigners();
-  const contractOwner = accounts[0];
-
-  // deploy DiamondCutFacet
-  const DiamondCutFacet = await ethers.getContractFactory("DiamondCutFacet");
-  const diamondCutFacet = await DiamondCutFacet.deploy();
-  await diamondCutFacet.deployed();
-
-  // deploy Diamond
-  const Diamond = await ethers.getContractFactory("Diamond");
-  const diamond = await Diamond.deploy(
-    contractOwner.address,
-    diamondCutFacet.address
-  );
-  await diamond.deployed();
-
-  // deploy DiamondInit
-  // DiamondInit provides a function that is called when the diamond is upgraded to initialize state variables
-  // Read about how the diamondCut function works here: https://eips.ethereum.org/EIPS/eip-2535#addingreplacingremoving-functions
-  const DiamondInit = await ethers.getContractFactory("DiamondInit");
-  const diamondInit = await DiamondInit.deploy();
-  await diamondInit.deployed();
-
-  // deploy facets
-  const FacetNames = ["DiamondLoupeFacet", "OwnershipFacet"];
-  const cut = [];
-  for (const FacetName of FacetNames) {
-    const Facet = await ethers.getContractFactory(FacetName);
-    const facet = await Facet.deploy();
-    await facet.deployed();
-    cut.push({
-      facetAddress: facet.address,
-      action: FacetCutAction.Add,
-      functionSelectors: getSelectors(facet),
-    });
-  }
-
-  // upgrade diamond with facets
-  const diamondCut = await ethers.getContractAt("IDiamondCut", diamond.address);
-  let tx;
-  let receipt;
-  // call to init function
-  let functionCall = diamondInit.interface.encodeFunctionData("init");
-  tx = await diamondCut.diamondCut(cut, diamondInit.address, functionCall);
-  receipt = await tx.wait();
-  if (!receipt.status) {
-    throw Error(`Diamond upgrade failed: ${tx.hash}`);
-  }
-  return diamond.address;
-}
-
-// We recommend this pattern to be able to use async/await everywhere
-// and properly handle errors.
-if (require.main === module) {
-  deployDiamond()
-    .then(() => process.exit(0))
-    .catch((error) => {
-      console.error(error);
-      process.exit(1);
-    });
-}
-
-exports.deployDiamond = deployDiamond;
