+++ types/index.ts
@@ -1,0 +1,9 @@
+export enum FacetCutAction {
+  Add,
+}
+
+export interface DiamondCut {
+  facetAddress: string;
+  action: FacetCutAction;
+  functionSelectors: string[];
+}
